Merge dev into main

+++ src/components/TicketCreate/TicketCreate.jsx
@@ -1,7 +1,8 @@
 import React, { useState, useEffect  } from "react";
-import { Link } from "react-router-dom";
+import { Link, createRoutesFromElements } from "react-router-dom";
 import { useSupabase } from '../../SupabaseContext';
 import './ticket.css';
+import Chat from '../Chat/Chat.jsx';
 
 
 const TicketCreation = () => {
@@ -13,40 +14,56 @@ const TicketCreation = () => {
         description: ""
     });
 
+    // currentUser data
+    const [currentUser, setCurrentUser] = useState('Guest')
+    
+
+    const [userId, setUserId] = useState(supabase.auth.getUser().then(user => {
+        setUserId(user.data.user.id)
+    }))
+
     const [showForm, setShowForm] = useState(false);
     // tickets in the Outstanding Ticket Status area
     const [pendingTickets, setPendingTickets] = useState([]);
     // state to select ticket by click
     const [selectedTicketIndex, setSelectedTicketIndex] = useState()
     // function to change state of form to show form or new issue button
-    const handleNewIssueClick = () => { setShowForm(true); }
+    const handleNewIssue = () => { setShowForm(true); }
     const handleCloseForm = () => { setShowForm(false); }
 
     // fetch all tickets from data base on inital load
     useEffect(() => {
         const fetchTickets = async () => {
-            const { data, error } = await supabase
-                .from('taskissue')
-                .select('*');
-
-            if (error) {
-                console.error('Error fetching tickets:', error.message);
-            } else {
-                setPendingTickets(data);
-                console.log(pendingTickets);
+            try {
+                const user = await supabase.auth.getUser()
+                const userId = user.data.user.id
+            
+                const { data, error } = await supabase
+                    .from('taskissue')
+                    .select('*')
+                    .eq('customer', userId)
+    
+                if (error) {
+                    console.error('Error fetching tickets:', error.message);
+                } else {
+                    setPendingTickets(data);
+                    console.log(pendingTickets);
+                } 
+            } catch (error) {
+                console.error('Error fetching tickets:', error.message)
             }
         };
-
+    
         fetchTickets();
-    }, []);
+    }, [supabase, userId]);
+    
 
     // add a issue to database and append to table 
     const handleSubmit = async (e) => {
         e.preventDefault();
-        console.log(supabase.auth.getUser())
     
         setPendingTickets([...pendingTickets, submittedTicket]);
-    
+        
         const { data, error } = await supabase
             .from('taskissue')
             .insert([
@@ -55,6 +72,7 @@ const TicketCreation = () => {
                     description: submittedTicket.description,
                     status: 'Pending',
                     category: submittedTicket.category,
+                    customer: userId, 
                 }
             ]);
     
@@ -82,84 +100,47 @@ const TicketCreation = () => {
         })
     }
 
-    // Delete Ticket from table and database when delete button is clicked
     const handleDeleteTicket = async (index) => {
-        // Retrieve the ticket ID from the local state
+
         const ticketIdToDelete = pendingTickets[index].id;
     
         try {
-            // Make an API call to delete the record from the Supabase table
-            const { data, error } = await supabase
+
+            const { error: messageDeleteError } = await supabase
+                .from('messages')
+                .delete()
+                .eq('ticket_id', ticketIdToDelete);
+    
+            if (messageDeleteError) {
+                console.error('Error deleting messages:', messageDeleteError.message);
+                return;
+            }
+    
+            const { error: ticketDeleteError } = await supabase
                 .from('taskissue')
                 .delete()
                 .eq('id', ticketIdToDelete);
     
-            if (error) {
-                console.error('Error deleting ticket:', error.message);
+            if (ticketDeleteError) {
+                console.error('Error deleting ticket:', ticketDeleteError.message);
                 return;
             }
     
-            // Update the local state to reflect the deletion
             setPendingTickets((prevTickets) =>
                 prevTickets.filter((ticket) => ticket.id !== ticketIdToDelete)
             );
         } catch (error) {
-            console.error('Error deleting ticket:', error.message);
+            console.error('Error during deletion process:', error.message);
         }
     };
     
 
     return (
-        <div className="ticketPageContainer">
-            <div className="ticketStatusContainer">
-                <h2>Outstanding Ticket Status</h2>
-                {/* Tickets table */}
-                {pendingTickets.length > 0 && (
-                    <table className="ticketsTable">
-                        <thead>
-                            <tr className="ticketHeaderContainer">
-                                <th className="headerCell">Category</th>
-                                <th className="locationCell">Location</th>
-                                <th className="descriptionCell">Description</th>
-                                <th className="actionCell">Action</th>
-                            </tr>
-                        </thead>
-                        <tbody className="ticketTableBody">
-                            {pendingTickets.map((ticket, index) => (
-                                <React.Fragment key={index}>
-                                    <tr key={index} onClick={() => setSelectedTicketIndex(index)}>
-                                        <td className="categoryCell">{ticket.category}</td>
-                                        <td className="locationCell">{ticket.location}</td>
-                                        <td className="descriptionCell, descriptionText">{ticket.description}</td>
-                                        <td className="actionCell">
-                                            <button onClick={() => handleDeleteTicket(index)}>Delete</button>
-                                        </td>
-                                    </tr>
-                                    {selectedTicketIndex === index && (
-                                        <tr key={index} className="expandedRowContainer">
-                                            <td colSpan="4">
-                                                <div className="expandedRow">
-                                                    <div className="fullDescription">&nbsp;<span className="ticketDownArrow">↳</span>&nbsp; {ticket.description}</div>
-                                                </div>
-                                            </td>
-                                        </tr>
-                                    )}
-
-                                </React.Fragment>
-                            ))}
-                        </tbody>
-                    </table>
-                )}
-                {/* Expanded row container outside the table */}
-                {/* {selectedTicketIndex !== undefined && (
-                    <div className="expandedRowContainer">
-                        <div className="expandedRow">
-                            <div className="fullDescription">{pendingTickets[selectedTicketIndex].description}</div>
-                        </div>
-                    </div>
-                )} */}
+        <div className="ticketPageContainer">    
+            <div id="currentUserContainer">
+                <p>Welcome, {currentUser}</p>
             </div>
-            {/* Show Form or new Issue button */}
+            {/* Show Form or Outstanding Tickets Status */}
             {showForm ? (
                 <div className="ticketFormContainer">
                     <div id='ticketFormClose' onClick={handleCloseForm}>x</div>
@@ -193,8 +174,8 @@ const TicketCreation = () => {
                             <option value="Austin">Austin, TX</option>
                             <option value="Dallas">Dallas, TX</option>
                             <option value="Houston">Houston, TX</option>
-                            <option value="SanAntonio">San Antonio, TX</option>
-                            <option value="LasVegas">Las Vegas, NV</option>
+                            <option value="San Antonio">San Antonio, TX</option>
+                            <option value="Las Vegas">Las Vegas, NV</option>
                             <option value="Carlsbad">Carlsbad, NM</option>
                             {/* Add more options as needed */}
                         </select>
@@ -212,7 +193,50 @@ const TicketCreation = () => {
                     </form>
                 </div>
             ) : (
-                <button onClick={handleNewIssueClick} id="newIssueButton">New Issue</button>
+                <>
+                    <div className="ticketStatusContainer">
+                    <h2>Outstanding Ticket Status</h2>
+                    {/* Tickets table */}
+                    {pendingTickets.length > 0 && (
+                        <table className="ticketsTable">
+                            <thead>
+                                <tr className="ticketHeaderContainer">
+                                    <th className="headerCell">Category</th>
+                                    <th className="locationCell">Location</th>
+                                    <th className="descriptionCell">Description</th>
+                                    <th className="actionCell">Action</th>
+                                </tr>
+                            </thead>
+                            <tbody className="ticketTableBody">
+                                {pendingTickets.map((ticket, index) => (
+                                    <React.Fragment key={index}>
+                                        <tr key={index} onClick={() => setSelectedTicketIndex(index)}>
+                                            <td className="categoryCell">{ticket.category}</td>
+                                            <td className="locationCell">{ticket.location}</td>
+                                            <td className="descriptionCell, descriptionText">{ticket.description}</td>
+                                            <td className="actionCell">
+                                                <button onClick={() => handleDeleteTicket(index)}>Delete</button>
+                                            </td>
+                                        </tr>
+                                        {selectedTicketIndex === index && (
+                                            <tr key={index} className="expandedRowContainer">
+                                                <td colSpan="4">
+                                                    <div className="expandedRow">
+                                                        <div className="fullDescription">&nbsp;<span className="ticketDescClose" onClick={() => setSelectedTicketIndex(null)}>X</span>  &nbsp;<span className="ticketDownArrow">↳</span>&nbsp; {ticket.description}</div>
+                                                    </div>
+                                                </td>
+                                                <Chat ticketId={ticket.id} />
+                                            </tr>
+                                        )}
+
+                                    </React.Fragment>
+                                ))}
+                            </tbody>
+                        </table>
+                    )}
+                </div>
+                <button onClick={handleNewIssue} id="newIssueButton">New Issue</button>
+                </>
             )}
             <div id="ticketLinkContainer">
                 <Link to='/'>Back Home</Link>
